Add concept of partial services
